🚚 Extract Environment class to gateway module
